refactor(ai): colocate provider options
